requirements.txt
python-telegram-bot==20.3
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
Pillow